+++ src/extension.ts
@@ -10,6 +10,7 @@ import * as status from './status'
 const enum Cached {
 	checksum = 'direnv.checksum',
 	environment = 'direnv.environment',
+	cwdOverride = 'direnv.cwdOverride',
 }
 type EnvCache = [string, string | undefined][]
 
@@ -26,6 +27,7 @@ class Direnv implements vscode.Disposable {
 	private viewBlocked = new vscode.EventEmitter<string>()
 	private didUpdate = new vscode.EventEmitter<void>()
 	private blockedPath?: string
+	private cwdOverride?: string
 	private watchers = vscode.Disposable.from()
 
 	constructor(private context: vscode.ExtensionContext, private status: status.Item) {
@@ -93,6 +95,36 @@ class Direnv implements vscode.Disposable {
 		this.didOpen(path)
 	}
 
+	async loadEnvrc(uri?: vscode.Uri) {
+		let envrcPath = ''
+		if (!uri) {
+			const options: vscode.OpenDialogOptions = {
+				canSelectFiles: true,
+				canSelectFolders: false,
+				canSelectMany: false,
+				defaultUri: vscode.Uri.file(
+					vscode.workspace.workspaceFolders?.[0].uri.path ?? process.cwd(),
+				),
+				openLabel: 'Load',
+				title: 'Select .envrc to load',
+			}
+			const uris = await vscode.window.showOpenDialog(options)
+
+			envrcPath = uris ? uris[0].path : ''
+		} else {
+			envrcPath = uri.path
+		}
+
+		if (envrcPath) {
+			if (envrcPath.substring(envrcPath.lastIndexOf('/') + 1) !== '.envrc') {
+				await vscode.window.showErrorMessage('direnv error: Not a .envrc!')
+				return
+			}
+			this.cwdOverride = envrcPath.substring(0, envrcPath.lastIndexOf('/'))
+			this.willLoad.fire()
+		}
+	}
+
 	async reload() {
 		await this.resetCache()
 		await this.load()
@@ -111,6 +143,7 @@ class Direnv implements vscode.Disposable {
 	}
 
 	private restoreCache(): Data | undefined {
+		this.cwdOverride = this.cache.get<string>(Cached.cwdOverride)
 		const checksum = this.cache.get<string>(Cached.checksum)
 		if (checksum === undefined) return
 		const entries = this.cache.get<EnvCache>(Cached.environment)
@@ -133,6 +166,7 @@ class Direnv implements vscode.Disposable {
 		}
 		await this.cache.update(Cached.checksum, hash.digest())
 		await this.cache.update(Cached.environment, entries)
+		await this.cache.update(Cached.cwdOverride, this.cwdOverride)
 	}
 
 	private async resetCache() {
@@ -220,7 +254,7 @@ class Direnv implements vscode.Disposable {
 		this.blockedPath = undefined
 		this.status.update(status.State.loading)
 		try {
-			const data = await direnv.dump()
+			const data = await direnv.dump(this.cwdOverride)
 			this.didLoad.fire(data)
 		} catch (err) {
 			if (err instanceof direnv.BlockedError) {
@@ -384,6 +418,12 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.commands.registerCommand(command.Direnv.open, async () => {
 			await instance.open()
 		}),
+		vscode.commands.registerCommand(
+			command.Direnv.loadEnvrc,
+			async (uri?: vscode.Uri) => {
+				await instance.loadEnvrc(uri)
+			},
+		),
 		vscode.window.tabGroups.onDidChangeTabs((e) => {
 			for (const tab of e.opened) {
 				if (tab.input instanceof vscode.TabInputText) {